backend/requirements.txt
fastapi
uvicorn[standard]
torch
transformers
sentence-transformers
faiss-cpu
reportlab